Extract FdReader to separate module
Before adding more input functions, extract FdReader to a separate
module to keep the input module clean.
